backend/requirements.txt
fastapi
uvicorn
scikit-learn
# GCP 
google-cloud-firestore
google-cloud-bigquery
google-cloud-aiplatform

# Utils
pydantic